backend/requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
amqp==5.0.2
apistar==0.7.2
apply-defaults==0.1.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
Babel==2.9.0
backcall==0.1.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.3.0
blessings==1.7
cachetools==4.1.0
celery==5.0.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
clipboard==0.0.4
cloudpickle==1.6.0
colorama==0.4.3
colorful==0.5.4
coolname==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
curlify==2.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.8
django-annoying==0.10.6
django-computed-property==0.3.0
django-countries==7.0
django-database-view==0.2.1
django-directory==0.0.3
django-email-verification==0.0.2
django-extensions==2.2.9
django-filter==2.4.0
django-haystack==3.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-language-field==0.0.3
django-mptt==0.11.0
django-pandas==0.6.2
django-polymorphic==2.1.2
django-recaptcha==2.0.6
djangorestframework==3.12.2
dm-tree==0.1.5
docstring-parser==0.7.3
docutils==0.16
drf-recaptcha==2.0.4
drf-spectacular==0.11.1
dukpy==0.2.3
easy-thumbnails==2.7
editdistance==0.5.3
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
flaky==3.7.0
flatbuffers==1.12
future==0.18.2
fuzzysearch==0.7.2
gast==0.3.3
geckodriver-autoinstaller==0.1.0
gin-config==0.4.0
gin-tune @ git+https://github.com/sergeivolodin/gin_tune.git@c87bcc432f8cd7bfe8ae652b9412733cbb081e0a
gitdb==4.0.5
GitPython==3.1.2
gpustat==0.6.0
graphql-core==3.1.2
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
hiredis==1.1.0
httplib2==0.19.0
hypothesis==5.41.4
hypothesis-graphql==0.3.2
hypothesis-jsonschema==0.18.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==5.2.1
ipympl==0.5.6
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.2.0
javascripthon==0.11
jedi==0.17.0
Jinja2==2.11.3
joblib==0.15.1
json5==0.9.4
jsonrpcclient==3.3.6
jsonrpcserver==4.1.3
jsonschema==3.2.0
junit-xml==1.9
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.3
jupyterlab-git==0.20.0
jupyterlab-server==1.1.5
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==5.0.2
langdetect==1.0.8
lazy-object-proxy==1.4.3
lxml==4.6.3
macropy3==1.1.0b2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
mplcursors==0.3
msgpack==1.0.0
multidict==4.7.6
namegenerator==1.0.6
names==0.3.0
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.6
nest-asyncio==1.3.3
networkx==2.4
notebook==6.1.5
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.11
opencensus-context==0.1.2
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
params-flow==0.8.2
parso==0.7.0
path==14.0.1
pathlib==1.0.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pip-chill==1.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.3
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
py-params==0.9.7
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.16.0
pytest==5.4.3
pytest-django==4.1.0
pytest-subtests==0.3.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.4
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
ray==1.0.1.post1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1
schemathesis==2.8.4
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
snowballstemmer==2.0.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sortedcontainers==2.3.0
soupsieve==2.0.1
Sphinx==3.3.1
sphinx-autodoc-annotation==1.0.post1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django==0.5.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
starlette==0.13.8
statsmodels==0.12.1
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
toml==0.10.2
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typesystem==0.2.4
typing-extensions==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yamlordereddictloader==0.4.0
yarl==1.6.3
youtube-dl==2021.4.1
zipp==3.1.0
